- Optimized main data loading - Receive screen fixes - Swap screen min receive and fee calculation corrections

+++ app/src/main/java/com/p2p/wallet/main/repository/MainRemoteRepository.kt
@@ -55,7 +55,7 @@ class MainRemoteRepository(
 
             return@withContext signatures
                 .map {
-                    async { getConfirmedTransaction(it.signature, it.slot.toLong()) }
+                    async { getConfirmedTransaction(it.signature, it.slot) }
                 }
                 .mapNotNull {
                     val response = it.await() ?: return@mapNotNull null
